+++ lib/cryptozoologist.rb
@@ -51,6 +51,11 @@ module Cryptozoologist
     Generator.city
   end
 
+  def self.full_name
+    Generator.full_name
+  end
+
+  protected
   def self.subdictionaries
     Dictionaries.library
   end