+++ lib/cryptozoologist/dictionaries.rb
@@ -31,6 +31,10 @@ module Cryptozoologist
         cities: {
           words: Cities::Words,
           terminologies: Cities::Terminologies
+        },
+        people: {
+          first_name: People::FirstName,
+          last_name: People::LastName,
         }
       }
     end
@@ -55,6 +59,13 @@ module Cryptozoologist
       create_list(:cities)
     end
 
+    def first_name
+      People::FirstName.list
+    end
+    
+    def last_name
+      People::LastName.list
+    end
     private
     def create_list(key)
       list = []